[Anoop] Added All controller for converting the value

+++ QuantityMeasurement_Web_API/Model/Kilogram.cs
@@ -16,7 +16,7 @@ namespace QuantityMeasurement_Web_API.Model
         /// Variable for kilogram value
         /// </summary>
         double kilogram;
-        public Kilogram(int kilogram)
+        public Kilogram(double kilogram)
         {
             this.kilogram = kilogram;
         }
